pyproject.toml
[project]
name = "django-celery-model"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.4.0",
    "django-admin-extra-buttons>=1.5.8",
    "django-celery-beat>=2.7.0",
    "django-concurrency>=2.5",
    "django>=4.2.16",
    "djangorestframework>=3.15.2",
    "redis>=5.0.8",
]

[tool.uv]
dev-dependencies = [
    "black>=24.8.0",
    "celery[pytest]>=5.4.0",
    "django-smart-env>=0.1.0",
    "django-webtest>=1.9.12",
    "factory-boy>=3.3.1",
    "flake8>=7.1.1",
    "isort>=5.13.2",
    "mkdocs-material>=9.5.38",
    "mkdocstrings[python]>=0.26.1",
    "mypy>=1.11.2",
    "pdbpp>=0.10.3",
    "pytest-cov>=5.0.0",
    "pytest-django>=4.9.0",
    "pytest>=8.3.3",
]



[tool.isort]
profile = "black"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
  | snapshots
)/
'''
